Update the frontend to the latest client library
