--- frontend/src/models/api/common.ts
@@ -1,6 +1,0 @@
-export interface HttpResponse<TStatus extends number, TData = undefined> {
-  status: TStatus;
-  data: TData;
-}
-
-export type UnknownHttpError = HttpResponse<0>;
