--- frontend/src/models/api/refresh.ts
@@ -1,4 +1,0 @@
-import { Login200ResponseData } from './login';
-
-// When refreshing an auth token, we get the same data as when logging in.
-export type AuthTokenRefresh200ResponseData = Login200ResponseData;
